Add concept of Framework Versioning
If we are a .NET 4.0 app, we are not compatible with .NET 4.5 DLLs even though
we're technically supposed to be. We need to specify at start what kind of app
we are so that we know to filter these out. As a bonus, also filter out .NET
Core DLLs and WinRT DLLs.